# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Single-user local Codex-like Agent API built with Go and CloudWeGo Eino.\nExternal effects always flow through ToolRouter, EffectInference, PolicyEngine, ApprovalCenter, and Executor.",
        "title": "Local Agent API",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/",
    "paths": {
        "/v1/approvals/pending": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Approvals"
                ],
                "summary": "List pending approvals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ApprovalListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/approvals/{approval_id}/approve": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Approvals"
                ],
                "summary": "Approve a pending action",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Approval ID",
                        "name": "approval_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ApprovalResolutionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/approvals/{approval_id}/reject": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Approvals"
                ],
                "summary": "Reject a pending action",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Approval ID",
                        "name": "approval_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rejection payload",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.RejectApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ApprovalResolutionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/kbs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Knowledge"
                ],
                "summary": "List knowledge bases",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.KnowledgeBaseListResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Knowledge"
                ],
                "summary": "Create knowledge base",
                "parameters": [
                    {
                        "description": "Knowledge base payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.CreateKnowledgeBaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.KnowledgeBaseItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/kbs/{kb_id}/documents/upload": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Knowledge"
                ],
                "summary": "Upload KB document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Knowledge base ID",
                        "name": "kb_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Document payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.KnowledgeBaseDocumentUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.KBChunkListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/kbs/{kb_id}/search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Knowledge"
                ],
                "summary": "Search knowledge base",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Knowledge base ID",
                        "name": "kb_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Search payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.KBChunkListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api_handlers.ApprovalListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_core.ApprovalRecord"
                    }
                }
            }
        },
        "internal_api_handlers.ApprovalResolutionResponse": {
            "type": "object",
            "properties": {
                "approval": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalRecord"
                },
                "result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                },
                "run": {
                    "$ref": "#/definitions/local-agent_internal_agent.RunResponse"
                }
            }
        },
...
        "internal_api_handlers.CreateKnowledgeBaseRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.KBChunkListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_core.KBChunk"
                    }
                }
            }
        },
        "internal_api_handlers.KnowledgeBaseDocumentUploadRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.KnowledgeBaseItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.KnowledgeBaseListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_core.KnowledgeBase"
                    }
                }
            }
        },
        "internal_api_handlers.LegacyErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.RejectApprovalRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.SearchRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "query": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_agent.AgentContext": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Message"
                    }
                }
            }
        },
        "local-agent_internal_agent.Plan": {
            "type": "object",
            "properties": {
                "decision": {
                    "$ref": "#/definitions/local-agent_internal_agent.PlanDecision"
                },
                "message": {
                    "type": "string"
                },
                "preamble": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "tool_proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                }
            }
        },
...
        "local-agent_internal_agent.RunResponse": {
            "type": "object",
            "properties": {
                "approval": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalRecord"
                },
                "assistant_message": {
                    "$ref": "#/definitions/local-agent_internal_core.Message"
                },
                "run_id": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/local-agent_internal_agent.RunState"
                },
                "tool_result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                }
            }
        },
        "local-agent_internal_agent.RunState": {
            "type": "object",
            "properties": {
                "approval_id": {
                    "type": "string"
                },
                "context": {
                    "$ref": "#/definitions/local-agent_internal_agent.AgentContext"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_step": {
                    "type": "string"
                },
                "current_step_index": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "max_steps": {
                    "type": "integer"
                },
                "plan": {
                    "$ref": "#/definitions/local-agent_internal_agent.Plan"
                },
                "policy": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                },
                "run_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/local-agent_internal_agent.RunStatus"
                },
                "step_count": {
                    "type": "integer"
                },
                "tool_result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_message": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_agent.RunStatus": {
            "type": "string",
            "enum": [
                "received_user_message",
                "context_built",
                "planned",
                "tool_proposed",
                "effect_inferred",
                "policy_decided",
                "approval_requested",
                "paused_for_approval",
                "approval_approved",
                "approval_rejected",
                "tool_executing",
                "tool_completed",
                "tool_failed",
                "assistant_summarizing",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "RunStatusReceived",
                "RunStatusContextBuilt",
                "RunStatusPlanned",
                "RunStatusToolProposed",
                "RunStatusEffectInferred",
                "RunStatusPolicyDecided",
                "RunStatusApprovalRequested",
                "RunStatusPausedForApproval",
                "RunStatusApprovalApproved",
                "RunStatusApprovalRejected",
                "RunStatusToolExecuting",
                "RunStatusToolCompleted",
                "RunStatusToolFailed",
                "RunStatusAssistantSummarize",
                "RunStatusCompleted",
                "RunStatusFailed",
                "RunStatusCancelled"
            ]
        },
...
        "local-agent_internal_core.ApprovalRecord": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "decision": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "id": {
                    "type": "string"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "input_snapshot": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                },
                "reason": {
                    "type": "string"
                },
                "resolved_at": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "snapshot_hash": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalStatus"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.ApprovalStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "ApprovalPending",
                "ApprovalApproved",
                "ApprovalRejected"
            ]
        },
...
        "local-agent_internal_core.EffectInferenceResult": {
            "type": "object",
            "properties": {
                "approval_required": {
                    "type": "boolean"
                },
                "confidence": {
                    "type": "number"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reason_summary": {
                    "type": "string"
                },
                "risk_level": {
                    "type": "string"
                },
                "sensitive": {
                    "type": "boolean"
                }
            }
        },
        "local-agent_internal_core.KBChunk": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kb_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "score": {
                    "type": "number"
                }
            }
        },
        "local-agent_internal_core.KnowledgeBase": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_core.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "content_json": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.PolicyDecision": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "approval_payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "reason": {
                    "type": "string"
                },
                "requires_approval": {
                    "type": "boolean"
                },
                "risk_level": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_core.ToolProposal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expected_effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "input": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "purpose": {
                    "type": "string"
                },
                "tool": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.ToolResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "output": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "started_at": {
                    "type": "string"
                },
                "tool_call_id": {
                    "type": "string"
                }
            }
        },
...
    }
}